requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
fastapi>=0.111.0
uvicorn>=0.30.0
python-docx>=1.1.0
markdown>=3.4.0

python-multipart>=0.0.9
